separating compute time
